Reduce vector profiling code bloat
If vector profiling is disabled at compile time, an instance of
vector_instrumentation is still created, and
vector_instrumentation::dump_max_size_histogram is still possibly
called. This causes the linker to keep a bunch of useless code around.

When vector profiling is disabled, disable uses of
vector_instrumentation to reduce the final binary size (if dead code
elimination is enabled when compiling and linking).

This commit should have no impact on tests.